Rename package
ovh.bailon.foodnet is already used in google play console
(by my closed account).
This renames the package to allows publishing the app on the playstore.
