--- app/src/main/java/ovh/bailon/foodnet/OnDataEventListener.java
@@ -1,9 +1,0 @@
-package ovh.bailon.foodnet;
-
-import java.util.ArrayList;
-
-public interface OnDataEventListener {
-    void onGetAllReady(ArrayList<OpenDating> list);
-
-    void onGetReady(OpenDating openDating);
-}
