--- app/src/main/java/ovh/bailon/foodnet/OnGroupEventListener.java
@@ -1,7 +1,0 @@
-package ovh.bailon.foodnet;
-
-import java.util.List;
-
-public interface OnGroupEventListener {
-    void onGetAllReady(List<String> list);
-}
