+++ app/src/main/java/ovh/bailon/foodnet2/FoodNetReceiver.java
@@ -1,0 +1,48 @@
+/*
+ * Copyright (C) 2020 Alexandre Bailon
+ *
+ * This program is free software; you can redistribute it and/or modify it under the terms of the
+ * GNU General Public License as published by the Free Software Foundation; either version 2 of
+ * the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
+ * PURPOSE. See the GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along with this program;
+ * If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package ovh.bailon.foodnet2;
+
+import android.app.Notification;
+import android.app.NotificationChannel;
+import android.app.NotificationManager;
+import android.content.BroadcastReceiver;
+import android.content.Context;
+import android.content.Intent;
+
+public class FoodNetReceiver extends BroadcastReceiver {
+    public static final String NOTIFICATION_CHANNEL_ID = "10001";
+    public static String NOTIFICATION_ID = "notification-id";
+    public static String NOTIFICATION = "notification";
+
+    public void onReceive(Context context, Intent intent) {
+        NotificationManager notificationManager =
+                (NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
+        Notification notification = intent.getParcelableExtra(NOTIFICATION);
+        if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.O) {
+            int importance = NotificationManager.IMPORTANCE_HIGH;
+            CharSequence name = context.getString(R.string.channel_name);
+            String description = context.getString(R.string.channel_description);
+            NotificationChannel notificationChannel =
+                    new NotificationChannel(NOTIFICATION_CHANNEL_ID, name, importance);
+            notificationChannel.setDescription(description);
+            assert notificationManager != null;
+            notificationManager.createNotificationChannel(notificationChannel);
+        }
+        int id = intent.getIntExtra(NOTIFICATION_ID, 0);
+        assert notificationManager != null;
+        notificationManager.notify(id, notification);
+    }
+}
