+++ app/src/main/java/ovh/bailon/foodnet2/OnDataEventListener.java
@@ -1,0 +1,9 @@
+package ovh.bailon.foodnet2;
+
+import java.util.ArrayList;
+
+public interface OnDataEventListener {
+    void onGetAllReady(ArrayList<OpenDating> list);
+
+    void onGetReady(OpenDating openDating);
+}
