+++ app/src/main/java/ovh/bailon/foodnet2/OnGroupEventListener.java
@@ -1,0 +1,7 @@
+package ovh.bailon.foodnet2;
+
+import java.util.List;
+
+public interface OnGroupEventListener {
+    void onGetAllReady(List<String> list);
+}
