+++ app/src/main/java/ovh/bailon/foodnet2/ui/ListItemAdapter.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright (C) 2020 Alexandre Bailon
+ *
+ * This program is free software; you can redistribute it and/or modify it under the terms of the
+ * GNU General Public License as published by the Free Software Foundation; either version 2 of
+ * the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
+ * PURPOSE. See the GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along with this program;
+ * If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package ovh.bailon.foodnet2.ui;
+
+import android.content.Context;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.ArrayAdapter;
+
+import java.util.ArrayList;
+
+public class ListItemAdapter extends ArrayAdapter<ListItemBase> {
+    public ListItemAdapter(Context context, ArrayList<ListItemBase> list) {
+        super(context, 0, list);
+    }
+
+    @Override
+    public View getView(int position, View convertView, ViewGroup parent) {
+        ListItemBase listItemBase = getItem(position);
+
+        return listItemBase.getView(position, convertView, parent);
+    }
+}
