+++ app/src/main/java/ovh/bailon/foodnet2/ui/SpinnerItemIcon.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright (C) 2020 Alexandre Bailon
+ *
+ * This program is free software; you can redistribute it and/or modify it under the terms of the
+ * GNU General Public License as published by the Free Software Foundation; either version 2 of
+ * the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
+ * PURPOSE. See the GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along with this program;
+ * If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package ovh.bailon.foodnet2.ui;
+
+public class SpinnerItemIcon {
+    int id;
+    int icon;
+    int text;
+
+    public SpinnerItemIcon(int id, int text, int icon) {
+        this.id = id;
+        this.text = text;
+        this.icon = icon;
+    }
+
+    public int getIcon() {
+        return icon;
+    }
+
+    public int getText() {
+        return text;
+    }
+
+    public int getId() {
+        return id;
+    }
+}
